pyproject.toml
[project]
name = "turismagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "faiss-cpu>=1.11.0.post1",
    "geoclip>=1.2.0",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "qwen-vl-utils>=0.0.11",
    "scipy>=1.16.1",
    "sentence-transformers>=5.1.0",
    "smolagents[gradio,litellm,mcp,telemetry,toolkit,transformers,vision]>=1.20.0",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "transformers-stream-generator>=0.0.5",
]
